finished

+++ remaining/scanner.cc
@@ -1429,7 +1429,7 @@ YY_RULE_SETUP
     yylloc.first_line = yylineno;
     yylloc.first_column = column;
     column += yyleng;
-    yylval.ival = atoi(yytext);
+    yylval.ival = std::stol(yytext);
     return T_INTNUM;
 }
 	YY_BREAK
